fix tests, move tests to integration and fix linter/formatting
