+++ yivi_app/integration_test/test_all.dart
@@ -11,6 +11,7 @@ import "embedded_issuance_flows/id_card_test.dart" as id_card_test;
 import "embedded_issuance_flows/passport_test.dart" as passport_test;
 import "embedded_issuance_flows/sms_issuance_test.dart" as sms_issuance_test;
 import "enroll_test.dart" as enroll_test;
+import "external_browser_credtypes_test.dart" as external_browser_test;
 import "issuance_test.dart" as issuance_test;
 import "issue_sdjwt_over_irma_test.dart" as irma_sdjwt_test;
 import "issue_wizard_test.dart" as issue_wizard_test;
@@ -24,6 +25,7 @@ import "settings_test.dart" as settings_test;
 
 /// Wrapper to execute all tests at once.
 void main() {
+  external_browser_test.main();
   irma_sdjwt_test.main();
   openid4vp_test.main();
   reorder_cards_test.main();
